+++ src/components/mainMenu/MenuButton.tsx
@@ -8,16 +8,16 @@ interface Props {
 }
 
 function MenuButton({ callback, icon, text }: Props) {
-  const handleClick = async (event: React.MouseEvent) => {
-    event.preventDefault();
-    await callback();
-  };
+  const handleClick = async () => await callback();
+
   return (
-    <a href="#" onClick={handleClick}>
-      <span className="text-slate-950 flex gap-2 text-5xl">
-        {icon} {text}
-      </span>
-    </a>
+    <span
+      className="text-slate-950 flex gap-2 text-5xl"
+      role="button"
+      onClick={handleClick}
+    >
+      {icon} {text}
+    </span>
   );
 }